Move Schema classe definitions from out_bigquery.rb
