+++ test/plugin/test_record_schema.rb
@@ -100,13 +100,13 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_load_schema
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, true)
     assert { fields.to_a.as_json == base_schema }
   end
 
   def test_load_schema_allow_overwrite_with_type_changed_column
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, true)
 
     fields.load_schema(base_schema_with_type_changed_column, true)
@@ -114,7 +114,7 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_load_schema_allow_overwrite_with_new_column
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, true)
 
     fields.load_schema(base_schema_with_new_column, true)
@@ -122,7 +122,7 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_load_schema_not_allow_overwrite_with_type_changed_column
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, false)
 
     fields.load_schema(base_schema_with_type_changed_column, false)
@@ -130,7 +130,7 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_load_schema_no_allow_overwrite_with_new_column
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, false)
 
     fields.load_schema(base_schema_with_new_column, false)
@@ -138,7 +138,7 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_format_one
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, false)
 
     time = Time.local(2016, 2, 7, 19, 0, 0).utc
@@ -155,7 +155,7 @@ class RecordSchemaTest < Test::Unit::TestCase
   end
 
   def test_format_one_with_extra_column
-    fields = Fluent::BigQueryOutput::RecordSchema.new("record")
+    fields = Fluent::BigQuery::RecordSchema.new("record")
     fields.load_schema(base_schema, false)
 
     time = Time.local(2016, 2, 7, 19, 0, 0).utc
